feat: integración completa de backend perecederos. Incluye recepción CEDIS, inventario por lote, auditoría de surtido, RPC, generación de etiquetas ZPL y ajustes en DTOs/servicios. Ademas se ajusta el dockerfile de backend y el docker-compose.yml

+++ Backend/src/main/java/com/bodegaaurrera/perecederos_demo/Controller/SurtidoController.java
@@ -8,6 +8,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
+
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("/api/surtido")
@@ -18,7 +20,7 @@ public class SurtidoController {
     @GetMapping("/surtido/{upc}/{cantidad}")
     public ResponseEntity<SugerenciaSurtidoDTO> sugerir(
             @PathVariable String upc,
-            @PathVariable int cantidad) {
+            @PathVariable BigDecimal cantidad) {
 
         return ResponseEntity.ok(
                 surtidoService.sugerirSurtido(upc, cantidad)
