+++ Backend/src/main/java/com/bodegaaurrera/perecederos_demo/DTO/DetallePlanoDTO.java
@@ -2,13 +2,14 @@ package com.bodegaaurrera.perecederos_demo.DTO;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
 public class DetallePlanoDTO {
     private Long idDetalleCedis;
     private String producto;          // nombre del producto
-    private Integer cantidadRecibida;
+    private BigDecimal cantidadRecibida;
     private String lote;
     private LocalDate fechaCaducidad;
     private int cantidadRpc;
